listing page refactor
