+++ styles/components/shared/_index.scss
@@ -13,3 +13,4 @@
 @forward 'selectedFilters';
 @forward 'banner';
 @forward 'slider';
+@forward 'toggle';
